centralize types, passthru clip options from user
