+++ packages/geoprocessing/src/dataproviders/index.ts
@@ -2,4 +2,4 @@
 export * from "./cog";
 export * from "./flatgeobuf";
 export * from "./getFeatures";
-export * from "./clipOperationLoader";
+export * from "./genClipOperationLoader";
