+++ packages/geoprocessing/src/toolbox/genClipToPolygonPreprocessor.ts
@@ -6,31 +6,7 @@ import area from "@turf/area";
 import { featureCollection as fc } from "@turf/helpers";
 import flatten from "@turf/flatten";
 import kinks from "@turf/kinks";
-
-/** Supported clip operations */
-export type ClipOperations = "intersect";
-
-/** Parameters for clip operation using polygon features */
-export interface FeatureClipOperation {
-  clipFeatures: Feature<Polygon>[];
-  operation: ClipOperations;
-}
-
-/** Parameters for clip operation using a datasource */
-export interface DatasourceClipOperation {
-  datasourceId: string;
-  operation: ClipOperations;
-}
-
-/** Optional parameters for preprocessor function */
-export interface ClipOptions {
-  /** Ensures result is a polygon. If clip results in multipolygon, returns the largest component */
-  ensurePolygon?: boolean;
-  /** maxSize in square kilometers that clipped polygon result can be.  Preprocessor function will throw if larger. */
-  maxSize?: number;
-  /** Whether or not maxSize should be enforced and throw */
-  enforceMaxSize?: boolean;
-}
+import { ClipOptions, FeatureClipOperation } from "../types/dataProcessor";
 
 /**
  * Returns an preprocessor function that uses opsLoad to fetch features
